Moved convert_params to the file caustics.parametrization so that it can be used more generally in other classes.
